MSVC Builds: Update "Installation"
This updates the glib-install.[props|vsprops] so that the headers to
"install" for the build are kept up-to-date.  Note that this has been
generated by a script, so that this hopefully means that the header
"installation" can be maintained along with the autotools files for most
uses, to simplify maintenance.