regex: Remove internal copy of libpcre
The internal copy of libpcre is at version 8.31, which is outdated, buggy
and vulnerable. Updating to a more recent version is difficult because
glib doesn't contain all the neccessary unicode data tables, see
[https://bugzilla.gnome.org/show_bug.cgi?id=689791] and
[https://bugzilla.gnome.org/show_bug.cgi?id=684222].

The reason for copying libpcre into glib was that it is patched to use
the glib unicode data tables instead of its own. However, it turns out
that libgio-2.0 links to libselinux, which in turn links to libpcre.
This means that all gtk+/gnome applications actually have two copies of
libpcre, more than negating any savings from the shared unicode tables.

Bump the required pcre version to the latest, which is 8.36.

Keep the --with-pcre configure switch which will be repurposed to select
between pcre and pcre2 soon.