GCancellable: add new poll wrapper APIs
Add two APIs to GCancellable for waiting on one or more GPollFD records until
they are ready, or until cancellation.

This is basically a cancellable version of g_poll().

This makes many typical operations easier; it will be possible to
replace all cases of g_cancellable_make_pollfd() in GIO with just the
simple version of this function, often at a large reduction in lines of
code.  Even better is that in the future, users of this API may be able
to avoid allocating an fd at all.

I'm not sure if it's worth keeping the _full version.

https://bugzilla.gnome.org/show_bug.cgi?id=741716