gio: use new g_cancellable_poll_simple() API
This removes a lot of similar-looking code.

This also means that there is no longer any code inside of GIO that is
handling fds from GCancellable.  That moves us closer to being able to
completely rid ourselves of per-cancellable file descriptors.

Windows parts are completely untested (and not even compiled).

https://bugzilla.gnome.org/show_bug.cgi?id=741716