Add g_main_context_ensure_is_owner()
This new function is intended to be used for verifying that operations
on objects that belong to a given main context are only made from valid
places.

The usual rule is that the operations must be performed while the main
context is acquired, but an exception is made for the global default
main context so that things can be set up before running the main loop.

https://bugzilla.gnome.org/show_bug.cgi?id=742599