Make GContextSpecificSource internally "public"
Expose GContextSpecificSource so that things in GIO can use it directly.

Add support for a user_data area in the struct which will be helpful for
some intended use cases.

Also expose a way to emit a signal directly on a single source without
having to go through the group.